fix(sonar): 修复可靠性 BUG 与安全热点，压制前端误报
消除空指针/类型转换/中断处理等 Java BUG；charts.ts ReDoS 分段解析；
sonar-project 压制 rpx、Lombok 字段、uni page 与 Vue @click a11y 误报。

+++ services/device-service/src/main/java/com/aicabinet/device/service/DeviceCommandTracker.java
@@ -184,7 +184,7 @@ public class DeviceCommandTracker {
         recent.entrySet().stream()
                 .sorted(Map.Entry.comparingByValue((a, b) ->
                         Long.compare(nullToZero(a.updatedAtMs()), nullToZero(b.updatedAtMs()))))
-                .limit(recent.size() - MAX_RECENT_COMMANDS)
+                .limit((long) recent.size() - MAX_RECENT_COMMANDS)
                 .map(Map.Entry::getKey)
                 .toList()
                 .forEach(recent::remove);
